requirements.txt

torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0
opencv-python>=4.7.0
gradio>=3.39.0
Pillow>=10.0.0
numpy>=1.24.0
